Moved AppEngineStandardProjectConfig back to  com.google.cloud.tools.eclipse.newproject.
