+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject.test/src/com/google/cloud/tools/eclipse/appengine/newproject/AppEngineStandardProjectConfigTest.java
@@ -11,8 +11,6 @@ import org.eclipse.core.runtime.NullProgressMonitor;
 import org.junit.Assert;
 import org.junit.Test;
 
-import com.google.cloud.tools.eclipse.util.templates.appengine.AppEngineStandardProjectConfig;
-
 public class AppEngineStandardProjectConfigTest {
 
   private AppEngineStandardProjectConfig config = new AppEngineStandardProjectConfig();
