+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject.test/src/com/google/cloud/tools/eclipse/appengine/newproject/CodeTemplatesTest.java
@@ -29,8 +29,6 @@ import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
-import com.google.cloud.tools.eclipse.util.templates.appengine.AppEngineStandardProjectConfig;
-
 public class CodeTemplatesTest {
 
   private SubMonitor monitor = SubMonitor.convert(new NullProgressMonitor());
