+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject.test/src/com/google/cloud/tools/eclipse/appengine/newproject/CreateAppEngineStandardWtpProjectTest.java
@@ -14,8 +14,6 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import com.google.cloud.tools.eclipse.util.templates.appengine.AppEngineStandardProjectConfig;
-
 import java.lang.reflect.InvocationTargetException;;
 
 @RunWith(MockitoJUnitRunner.class)
