+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/src/com/google/cloud/tools/eclipse/appengine/newproject/AppEngineStandardProjectConfig.java
@@ -1,16 +1,13 @@
-package com.google.cloud.tools.eclipse.util.templates.appengine;
+package com.google.cloud.tools.eclipse.appengine.newproject;
 
 import org.eclipse.core.resources.IProject;
-
 import java.io.File;
 import java.net.URI;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * Collects all data needed to create and configure an App Engine Standard Project.
  */
-public class AppEngineStandardProjectConfig extends TemplateDataModel {
+class AppEngineStandardProjectConfig {
   private File cloudSdkLocation = null;
   private URI eclipseProjectLocationUri = null;
   private String appEngineProjectId = "";
@@ -60,13 +57,6 @@ public class AppEngineStandardProjectConfig extends TemplateDataModel {
 
   public void setEclipseProjectLocationUri(URI uri) {
     this.eclipseProjectLocationUri = uri;
-  }
-
-  @Override
-  public Map<String, Object> getDataMap() {
-    Map<String, Object> root = new HashMap<>();
-    root.put("project", this);
-    return root;
   }
 
 }
