+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/src/com/google/cloud/tools/eclipse/appengine/newproject/CodeTemplates.java
@@ -22,8 +22,6 @@ import org.eclipse.core.runtime.Path;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.SubMonitor;
 
-import com.google.cloud.tools.eclipse.util.templates.appengine.AppEngineStandardProjectConfig;
-
 public class CodeTemplates {
 
   /**
